Spring Boot Practice. Order 2.0 replaced logic OrderItemService(Controller) to OrderService deleted Status entity and replace enum Status to Order entity divided logic to separate methods -> placeOrder(...)
