+++ src/main/java/shop/controller/BookController.java
@@ -37,7 +37,7 @@ public class BookController {
             summary = "Create a new book",
             description = "Create a new book"
     )
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    //@PreAuthorize("hasRole('ROLE_ADMIN')")
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping
     public BookDto createBook(@RequestBody @Valid CreateBookRequestDto createBookDto) {
@@ -68,7 +68,7 @@ public class BookController {
             summary = "Update a book by id",
             description = "Update a specific book"
     )
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    //@PreAuthorize("hasRole('ROLE_ADMIN')")
     @PutMapping("/{id}")
     public BookDto updateBook(@PathVariable @Positive Long id,
                               @RequestBody @Valid UpdateBookRequestDto updateBookDto) {
@@ -79,7 +79,7 @@ public class BookController {
             summary = "Delete a book by id",
             description = "Delete a specific book"
     )
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    //@PreAuthorize("hasRole('ROLE_ADMIN')")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/{id}")
     public void deleteBookById(@PathVariable @Positive Long id) {
