+++ src/main/java/shop/controller/CategoryController.java
@@ -31,7 +31,7 @@ public class CategoryController {
     @Operation(
             summary = "Create a new category",
             description = "Create a new category")
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    //@PreAuthorize("hasRole('ROLE_ADMIN')")
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping
     public CategoryDto createCategory(@RequestBody @Valid CreateCategoryRequestDto requestDto) {
@@ -62,7 +62,7 @@ public class CategoryController {
             summary = "Update a category by id",
             description = "Update a specific category"
     )
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    //@PreAuthorize("hasRole('ROLE_ADMIN')")
     @PostMapping("/{id}")
     public CategoryDto updateCategory(@PathVariable @Positive Long id,
                                       @RequestBody @Valid UpdateCategoryRequestDto requestDto) {
@@ -73,7 +73,7 @@ public class CategoryController {
             summary = "Delete a category by id",
             description = "Delete a specific category"
     )
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    //@PreAuthorize("hasRole('ROLE_ADMIN')")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/{id}")
     public void deleteCategory(@PathVariable @Positive Long id) {
