+++ src/main/java/shop/controller/OrderController.java
@@ -5,6 +5,8 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Positive;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.MediaType;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
@@ -17,6 +19,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import shop.dto.request.order.PlaceOrderRequestDto;
 import shop.dto.request.order.StatusDto;
+import shop.dto.responce.order.OrderItemDto;
 import shop.dto.responce.order.OrderResponseDto;
 import shop.service.OrderService;
 
@@ -52,10 +55,34 @@ public class OrderController {
             summary = "Update order status",
             description = "Update order status"
     )
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    //@PreAuthorize("hasRole('ROLE_ADMIN')")
     @PatchMapping("/{id}")
     public OrderResponseDto updateOrderStatus(@PathVariable @Positive Long id,
                                               @RequestBody @Valid StatusDto statusDto) {
         return orderService.updateOrderStatus(id, statusDto);
+    }
+
+    @Operation(
+            summary = "Get all order items",
+            description = "Retrieve all order items for a specific order"
+    )
+    @PreAuthorize("hasRole('ROLE_USER')")
+    @GetMapping("/{orderId}/items")
+    public Page<OrderItemDto> getOrderItems(Authentication authentication,
+                                            @PathVariable @Positive Long orderId,
+                                            Pageable pageable) {
+        return orderService.getOrderItems(authentication, orderId, pageable);
+    }
+
+    @Operation(
+            summary = "Get a order item",
+            description = "Retrieve a specific order item within an order"
+    )
+    @PreAuthorize("hasRole('ROLE_USER')")
+    @GetMapping("/{orderId}/items/{itemId}")
+    public OrderItemDto getOrderItem(Authentication authentication,
+                                     @PathVariable @Positive Long orderId,
+                                     @PathVariable @Positive Long itemId) {
+        return orderService.getOrderItem(authentication, orderId, itemId);
     }
 }
