--- src/main/java/shop/entity/Status.java
@@ -1,31 +1,0 @@
-package shop.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Table(name = "statuses")
-@Getter
-@Setter
-public class Status {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private StatusName status;
-
-    public enum StatusName {
-        COMPLETED,
-        PENDING,
-        DELIVERED
-    }
-}
