+++ src/main/java/shop/entity/User.java
@@ -18,7 +18,6 @@ import lombok.Setter;
 import lombok.ToString;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.Where;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
@@ -81,10 +80,5 @@ public class User implements UserDetails {
     @Override
     public boolean isEnabled() {
         return !isDeleted;
-    }
-
-    public static Long getUserIdFromAuthentication(Authentication authentication) {
-        User user = (User) authentication.getPrincipal();
-        return user.getId();
     }
 }
