+++ src/main/java/shop/repository/OrderRepository.java
@@ -7,5 +7,8 @@ import shop.entity.Order;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
     @EntityGraph(attributePaths = {"orderItems", "status"})
+    Optional<Order> findOrderById(Long id);
+
+    @EntityGraph(attributePaths = {"orderItems", "status"})
     Optional<Order> findOrderByUserId(Long userId);
 }
