+++ src/main/java/shop/security/CustomUserDetailsService.java
@@ -1,10 +1,12 @@
 package shop.security;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+import shop.entity.User;
 import shop.exception.EntityNotFoundException;
 import shop.repository.UserRepository;
 
@@ -17,5 +19,10 @@ public class CustomUserDetailsService implements UserDetailsService {
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
         return userRepository.findUserByEmail(username)
                 .orElseThrow(() -> new EntityNotFoundException("User " + username + " not found"));
+    }
+
+    public static Long getUserIdFromAuthentication(Authentication authentication) {
+        User user = (User) authentication.getPrincipal();
+        return user.getId();
     }
 }
