+++ src/main/java/shop/service/OrderService.java
@@ -1,8 +1,11 @@
 package shop.service;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.security.core.Authentication;
 import shop.dto.request.order.PlaceOrderRequestDto;
 import shop.dto.request.order.StatusDto;
+import shop.dto.responce.order.OrderItemDto;
 import shop.dto.responce.order.OrderResponseDto;
 
 public interface OrderService {
@@ -11,4 +14,9 @@ public interface OrderService {
     OrderResponseDto getOrderHistory(Authentication authentication);
 
     OrderResponseDto updateOrderStatus(Long id, StatusDto statusDto);
+
+    Page<OrderItemDto> getOrderItems(Authentication authentication,
+                                     Long orderId, Pageable pageable);
+
+    OrderItemDto getOrderItem(Authentication authentication, Long orderId, Long itemId);
 }
