+++ src/main/java/shop/service/impl/OrderServiceImpl.java
@@ -1,31 +1,34 @@
 package shop.service.impl;
 
-import static shop.entity.User.getUserIdFromAuthentication;
+import static shop.security.CustomUserDetailsService.getUserIdFromAuthentication;
 
 import jakarta.transaction.Transactional;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import shop.dto.request.order.PlaceOrderRequestDto;
 import shop.dto.request.order.StatusDto;
 import shop.dto.responce.cart.item.CartItemResponseDto;
+import shop.dto.responce.order.OrderItemDto;
 import shop.dto.responce.order.OrderResponseDto;
 import shop.dto.responce.shopping.cart.ShoppingCartResponseDto;
 import shop.entity.Book;
 import shop.entity.Order;
 import shop.entity.OrderItem;
-import shop.entity.Status;
 import shop.entity.User;
 import shop.exception.EntityNotFoundException;
+import shop.mapper.OrderItemMapper;
 import shop.mapper.OrderMapper;
 import shop.repository.BookRepository;
 import shop.repository.OrderItemRepository;
 import shop.repository.OrderRepository;
-import shop.repository.StatusRepository;
 import shop.security.CustomUserDetailsService;
 import shop.service.OrderService;
 import shop.service.ShoppingCartService;
@@ -33,28 +36,80 @@ import shop.service.ShoppingCartService;
 @Service
 @RequiredArgsConstructor
 public class OrderServiceImpl implements OrderService {
+    private final CustomUserDetailsService userDetailsService;
     private final ShoppingCartService shoppingCartService;
     private final BookRepository bookRepository;
-    private final CustomUserDetailsService userDetailsService;
     private final OrderItemRepository orderItemRepository;
     private final OrderRepository orderRepository;
-    private final StatusRepository statusRepository;
     private final OrderMapper orderMapper;
+    private final OrderItemMapper orderItemMapper;
 
     @Transactional
     @Override
     public OrderResponseDto placeOrder(Authentication authentication,
                                        PlaceOrderRequestDto requestDto) {
-        User user = (User)userDetailsService.loadUserByUsername(authentication.getName());
+        User user = (User) userDetailsService.loadUserByUsername(authentication.getName());
+        Order order = createOrder(user, requestDto);
+        Set<OrderItem> orderItems = createOrderItems(authentication, order);
+        order.setOrderItems(orderItems);
+        order.setTotal(getTotal(orderItems));
+        orderRepository.save(order);
+        orderItemRepository.saveAll(orderItems);
+        return orderMapper.toOrderResponseDto(order);
+    }
+
+    @Override
+    public OrderResponseDto getOrderHistory(Authentication authentication) {
+        Long userId = getUserIdFromAuthentication(authentication);
+        return orderMapper.toOrderResponseDto(
+                orderRepository.findOrderByUserId(userId)
+                        .orElseThrow(() -> new EntityNotFoundException("Can`t find order: "
+                                + userId)));
+    }
+
+    @Override
+    public OrderResponseDto updateOrderStatus(Long id, StatusDto statusDto) {
+        Order order = orderRepository.findOrderById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Can`t find order by id: "
+                        + id));
+        order.setStatus(Order.Status.valueOf(statusDto.getStatus()));
+        orderRepository.save(order);
+        return orderMapper.toOrderResponseDto(order);
+    }
+
+    @Override
+    public Page<OrderItemDto> getOrderItems(Authentication authentication,
+                                            Long orderId, Pageable pageable) {
+        Long userId = getUserIdFromAuthentication(authentication);
+        Page<OrderItem> orderItems = orderItemRepository
+                .findByOrderIdAndOrderUserId(orderId, userId, pageable);
+        return orderItems.map(orderItemMapper::toDto);
+    }
+
+    @Override
+    public OrderItemDto getOrderItem(Authentication authentication, Long orderId, Long itemId) {
+        Long userId = getUserIdFromAuthentication(authentication);
+        List<OrderItem> orderItems = orderItemRepository
+                .findByOrderIdAndOrderUserId(userId, orderId);
+        OrderItem orderItem = orderItems.stream()
+                .filter(item -> item.getId().equals(itemId))
+                .findFirst()
+                .orElseThrow(() -> new EntityNotFoundException("Can't find OrderItem by id "
+                        + itemId));
+        return orderItemMapper.toDto(orderItem);
+    }
+
+    private Order createOrder(User user, PlaceOrderRequestDto requestDto) {
         Order order = new Order();
         order.setUser(user);
-        order.setStatus(statusRepository.findByStatus(Status.StatusName.COMPLETED)
-                .orElseThrow(() -> new EntityNotFoundException("Can`t find status: "
-                        + Status.StatusName.COMPLETED)));
+        order.setStatus(Order.Status.COMPLETED);
         order.setOrderDate(LocalDateTime.now());
         order.setShippingAddress(requestDto.getShippingAddress());
+        return order;
+    }
+
+    private Set<OrderItem> createOrderItems(Authentication authentication, Order order) {
         Set<OrderItem> orderItems = new HashSet<>();
-        BigDecimal total = BigDecimal.ZERO;
         ShoppingCartResponseDto shoppingCart = shoppingCartService
                 .getShoppingCartById(authentication);
         Set<CartItemResponseDto> cartItems = shoppingCart.getCartItems();
@@ -70,35 +125,13 @@ public class OrderServiceImpl implements OrderService {
             orderItem.setPrice(price.multiply(BigDecimal.valueOf(
                     cartItem.getQuantity())));
             orderItems.add(orderItem);
-            total = total.add(orderItem.getPrice());
         }
-        order.setOrderItems(orderItems);
-        order.setTotal(total);
-        orderRepository.save(order);
-        orderItemRepository.saveAll(orderItems);
-        return orderMapper.toOrderResponseDto(order);
+        return orderItems;
     }
 
-    @Override
-    public OrderResponseDto getOrderHistory(Authentication authentication) {
-        Long userId = getUserIdFromAuthentication(authentication);
-        return orderMapper.toOrderResponseDto(
-                orderRepository.findOrderByUserId(userId)
-                        .orElseThrow(() -> new EntityNotFoundException("Can`t find order: "
-                                + userId)));
-    }
-
-    @Transactional
-    @Override
-    public OrderResponseDto updateOrderStatus(Long id, StatusDto statusDto) {
-        Order order = orderRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Can`t find order by id: "
-                        + id));
-        order.setStatus(statusRepository.findByStatus(
-                        Status.StatusName.valueOf(statusDto.getStatus()))
-                        .orElseThrow(() -> new EntityNotFoundException("Can`t find status:"
-                                + statusDto)));
-        orderRepository.save(order);
-        return orderMapper.toOrderResponseDto(order);
+    private BigDecimal getTotal(Set<OrderItem> orderItems) {
+        return orderItems.stream()
+                .map(OrderItem::getPrice)
+                .reduce(BigDecimal.ZERO, BigDecimal::add);
     }
 }
