--- src/main/resources/db/changelog/changes/12-insert-statuses.yaml
@@ -1,23 +1,0 @@
-databaseChangeLog:
-  - changeSet:
-      id: insert-statuses
-      author: dorofieieva
-      changes:
-        - insert:
-            tableName: statuses
-            columns:
-              - column:
-                  name: status
-                  value: 'COMPLETED'
-        - insert:
-            tableName: statuses
-            columns:
-              - column:
-                  name: status
-                  value: 'PENDING'
-        - insert:
-            tableName: statuses
-            columns:
-              - column:
-                  name: status
-                  value: 'DELIVERED'
