+++ src/main/resources/db/changelog/db.changelog-master.yaml
@@ -20,12 +20,8 @@ databaseChangeLog:
   - include:
       file: db/changelog/changes/10-create-carts-items-table.yaml
   - include:
-      file: db/changelog/changes/11-create-statuses-table.yaml
+      file: db/changelog/changes/11-create-orders-table.yaml
   - include:
-      file: db/changelog/changes/12-insert-statuses.yaml
+      file: db/changelog/changes/12-create-order-items-table.yaml
   - include:
-      file: db/changelog/changes/13-create-orders-table.yaml
-  - include:
-      file: db/changelog/changes/14-create-order-items-table.yaml
-  - include:
-      file: db/changelog/changes/15-create-orders-items-table.yaml
+      file: db/changelog/changes/13-create-orders-items-table.yaml
